requirements.txt
pysodium==0.7.17
pycryptodome==3.20.0